Adding component that shows all polls
