+++ src/components/Main.js
@@ -1,7 +1,8 @@
 import React from "react";
-import { Switch, Route } from "react-router-dom";
+import { Switch, Route, Link, } from "react-router-dom";
 import NewPoll from "./NewPoll/NewPoll";
 import Poll from "./Poll/Poll";
+import Polls from "./Polls/Polls";
 
 const Main = () => (
   <div className="main" role="main">
@@ -9,7 +10,8 @@ const Main = () => (
       <Switch>
         <Route exact path="/" component={NewPoll} />
         <Route path="/new-poll" component={NewPoll} />
-        <Route path="/poll/:pollId" component={Poll} />
+        <Route path="/polls" component={Polls} />
+        <Route path="/poll/:pollId" component={Poll} />>
       </Switch>
     </div>
   </div>
