+++ src/components/NewPoll/NewPoll.js
@@ -11,7 +11,6 @@ import {
   handleChangePollGroup,
   handleChangePollQuestion,
   handleFormSubmit,
-  getPollQuestions,
 } from "./NewPoll.actions";
 
 const StyledPoll = styled.div`
@@ -21,7 +20,6 @@ const StyledPoll = styled.div`
 class NewPoll extends Component {
   componentDidMount() {
     this.props.fetchPollGroups();
-    this.props.getPollQuestions();
   }
 
   render() {
@@ -74,7 +72,6 @@ const mapDispatchToProps = {
   handleChangePollGroup,
   handleChangePollQuestion,
   handleFormSubmit,
-  getPollQuestions,
 };
 
 export default connect(
