+++ src/index.js
@@ -10,10 +10,13 @@ import registerServiceWorker from './registerServiceWorker';
 import { store } from "./store";
 
 ReactDOM.render(
-    <Provider store={store}>
-      <BrowserRouter>
-        <App />
-      </BrowserRouter>
-    </Provider>,
-document.getElementById('root'));
+  <Provider store={store}>
+    <BrowserRouter>
+      
+      <App />
+
+    </BrowserRouter>
+  </Provider>,
+  document.getElementById("root")
+);
 registerServiceWorker();
